pyproject.toml
[tool.poetry]
name = "nbmf_reverse"
version = "0.1.0"
description = ""
authors = [""]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
matplotlib = "^3.9.2"
jupyter = "^1.1.1"
SciencePlots = "^2.1.1"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
tqdm = "^4.66.6"
dwave-ocean-sdk = "^8.2.0"
joblib = "^1.4.2"
scipy = "^1.15.2"
pandas = "^2.2.3"
gurobipy = "^12.0.1"

[tool.poetry.dev-dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
